# Speed improvements in WKB geometry import, and ogr2ogr