--- a/ogr/ogrmultilinestring.cpp
+++ b/ogr/ogrmultilinestring.cpp
@@ -138,6 +138,46 @@ OGRMultiLineString::isCompatibleSubType(OGRwkbGeometryType eGeomType) const
 }
 
 /************************************************************************/
+/*                           importFromWkb()                            */
+/************************************************************************/
+
+OGRErr OGRMultiLineString::importFromWkb(const unsigned char *pabyData,
+                                         size_t nSize,
+                                         OGRwkbVariant eWkbVariant,
+                                         size_t &nBytesConsumedOut)
+
+{
+    if (nGeomCount == 1 && nSize >= 9 && flags == 0 && pabyData[0] == wkbNDR &&
+        memcmp(pabyData + 1, "\x05\x00\x00\x00\x01\x00\x00\x00", 8) == 0)
+    {
+        // Optimization to import a Intel-ordered 1-part multilinestring on
+        // top of an existing 1-part multilinestring, to save dynamic memory
+        // allocations.
+        const size_t nDataOffset = 9;
+        size_t nBytesConsumedLineString = 0;
+        // cppcheck-suppress knownConditionTrueFalse
+        if (nSize != static_cast<size_t>(-1))
+            nSize -= nDataOffset;
+        OGRErr eErr = cpl::down_cast<OGRLineString *>(papoGeoms[0])
+                          ->OGRLineString::importFromWkb(
+                              pabyData + nDataOffset, nSize, eWkbVariant,
+                              nBytesConsumedLineString);
+        if (eErr == OGRERR_NONE)
+        {
+            nBytesConsumedOut = nDataOffset + nBytesConsumedLineString;
+        }
+        else
+        {
+            empty();
+        }
+        return eErr;
+    }
+
+    return OGRGeometryCollection::importFromWkbInternal(
+        pabyData, nSize, /*nRecLevel=*/0, eWkbVariant, nBytesConsumedOut);
+}
+
+/************************************************************************/
 /*                            exportToWkt()                             */
 /************************************************************************/