--- a/ogr/ogrpolygon.cpp
+++ b/ogr/ogrpolygon.cpp
@@ -351,9 +351,39 @@ OGRErr OGRPolygon::importFromWkb(const unsigned char *pabyData, size_t nSize,
                                  size_t &nBytesConsumedOut)
 
 {
-    nBytesConsumedOut = 0;
     OGRwkbByteOrder eByteOrder = wkbNDR;
     size_t nDataOffset = 0;
+
+    if (oCC.nCurveCount == 1 && flags == 0 && nSize >= 9 &&
+        pabyData[0] == wkbNDR &&
+        memcmp(pabyData + 1, "\x03\x00\x00\x00\x01\x00\x00\x00", 8) == 0)
+    {
+        // Optimization to import a Intel-ordered 1-ring polygon on
+        // top of an existing 1-ring polygon, to save dynamic memory
+        // allocations.
+        size_t nBytesConsumedRing = 0;
+        nDataOffset = 9;
+        // cppcheck-suppress knownConditionTrueFalse
+        if (nSize != static_cast<size_t>(-1))
+            nSize -= nDataOffset;
+        OGRErr eErr =
+            cpl::down_cast<OGRLinearRing *>(oCC.papoCurves[0])
+                ->_importFromWkb(eByteOrder, flags, pabyData + nDataOffset,
+                                 nSize, nBytesConsumedRing);
+        if (eErr == OGRERR_NONE)
+        {
+            nBytesConsumedOut = nDataOffset + nBytesConsumedRing;
+        }
+        else
+        {
+            empty();
+        }
+
+        return eErr;
+    }
+
+    nBytesConsumedOut = 0;
+
     // coverity[tainted_data]
     OGRErr eErr = oCC.importPreambleFromWkb(this, pabyData, nSize, nDataOffset,
                                             eByteOrder, 4, eWkbVariant);