… ogr/ogrlinestring.cpp
… ogr/ogrsf_frmts/generic/ogrlayerarrow.cpp