*   OGRLineString/Polygon/MultiPolygon/MultiLineString: make it possible to run importFromWkb() on the same object and limiting the number of dynamic memory (re)allocations
    
    - For lines, we keep track of the maximum capacity of the point array,
      with a new member variable.
      So doing setNumPoints(10), then setNumPoints(3), then setNumPoints(9)
      just allocates an array of size 10.
    - For polygons, if importing a WKB for a single ring polygon on top of
      an existing single ring polygon, reuse the existing ring (and benefit
      from the previous optimization)
    - For multipolygons, if importing a WKB for a single part multipolygon
      on top of an existing single part multipolygon, reuse the existing
      polygon (and benefit from the previous optimization)
    - Similar for multilinestring

* WriteArrowBatch() generic implementation: reuse existing geometry object to save memory allocations

* ogr2ogr: use Arrow interface as soon as the input driver declares OLCFastGetArrowStream, even if the output driver doesn't declare OLCFastWriteArrowBatch
    
    This helps in a notable way for Parquet -> GPKG or GPKG -> GPKG when
    disabling the creation of the spatial index which is now the major time
    consumer.
    
    Now (using Arrow API):
    
    ```
    $ time  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress
    0...10...20...30...40...50...60...70...80...90...100 - done.
    
    real    0m12,868s
    user    0m13,338s
    sys     0m1,843s
    ```
    
    Without use of Arrow API:
    
    ```
    $ time  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config OGR2OGR_USE_ARROW_API NO
    0...10...20...30...40...50...60...70...80...90...100 - done.
    
    real    0m17,625s
    user    0m15,917s
    sys     0m1,704s
    ```