@tbonfort 2023-10-13
nice speedup! could you give the before/after outputs of /usr/bin/time instead of the builtin time in order to also show memory consumption evolution?