@tbonfort 2023-10-13
my apologies Even, rewording my comment as it was not my intension to be pushy:
fyi, I believe that using /usr/bin/time instead of the builtin time function would also print out memory usage information
regards 😀