@rouault 2023-10-13
@tbonfort  Very relevant question, that made me realized the  fix of  fcbed4e6bd was needed ....

Without the optimization using the GetNextFeature() + CreateFeature() looping strategy
$ /usr/bin/time -v  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config OGR2OGR_USE_ARROW_API NO
0...10...20...30...40...50...60...70...80...90...100 - done.
	Command being timed: "ogr2ogr out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config OGR2OGR_USE_ARROW_API NO"
	User time (seconds): 18.56
	System time (seconds): 1.78
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:20.38
	Maximum resident set size (kbytes): 170408

With optimization, using the default 4-thread strategy in the GeoPackage driver, to prefetch the input data (ie the GeoPackage driver not only acquires a single batch of feature when requested but spawns threads to fetch the next ones):
$ /usr/bin/time -v  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress
0...10...20...30...40...50...60...70...80...90...100 - done.
	Command being timed: "ogr2ogr out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress"
	User time (seconds): 14.38
	System time (seconds): 1.92
	Percent of CPU this job got: 117%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:13.90
	Maximum resident set size (kbytes): 656096

With optimization, prefetching limited to 1 thread:
$ /usr/bin/time -v  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config GDAL_NUM_THREADS 1
0...10...20...30...40...50...60...70...80...90...100 - done.
	Command being timed: "ogr2ogr out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config GDAL_NUM_THREADS 1"
	User time (seconds): 14.29
	System time (seconds): 1.50
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:15.81
	Maximum resident set size (kbytes): 332876

The amount of memory is indeed larger with the use of the Arrow API, which is expected. By default we acquire batches of a number of features up to the value of the -gt parameter, which defaults to 100,000.

When limiting to 10,000:

$ /usr/bin/time -v  ogr2ogr  out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config GDAL_NUM_THREADS 1 -gt 10000
0...10...20...30...40...50...60...70...80...90...100 - done.
	Command being timed: "ogr2ogr out.gpkg nz-building-outlines.gpkg -lco spatial_index=no -progress --config GDAL_NUM_THREADS 1 -gt 10000"
	User time (seconds): 14.39
	System time (seconds): 1.56
	Percent of CPU this job got: 99%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:15.96
	Maximum resident set size (kbytes): 497184

I cannot really make sense of that last result. Why would decreasing the size of the batch increase the RAM usage... ?